+++ RiskOfChaos/ModifierController/Pickups/IPickupModificationProvider.cs
@@ -1,0 +1,6 @@
+﻿namespace RiskOfChaos.ModifierController.Pickups
+{
+    public interface IPickupModificationProvider : IValueModificationProvider<PickupModificationInfo>
+    {
+    }
+}