+++ RiskOfChaos/ModifierController/Pickups/PickupModificationInfo.cs
@@ -1,0 +1,11 @@
+﻿namespace RiskOfChaos.ModifierController.Pickups
+{
+    public struct PickupModificationInfo
+    {
+        public uint BounceCount = 0;
+
+        public PickupModificationInfo()
+        {
+        }
+    }
+}